+++ adapters/Moloco/MolocoAdapter/MolocoUtils.swift
@@ -27,10 +27,10 @@ final class MolocoUtils {
   }
 
   static func log(_ logMessage: String) {
-    print("GADMediationAdapterMoloco - \(logMessage)")
+    print("MediationAdapterMoloco - \(logMessage)")
   }
 
-  static func getAdUnitId(from adConfiguration: GADMediationAdConfiguration) -> String? {
+  static func getAdUnitId(from adConfiguration: MediationAdConfiguration) -> String? {
     adConfiguration.isTestRequest
       ? MolocoConstants.molocoTestAdUnitName
       : adConfiguration.credentials.settings[MolocoConstants.adUnitIdKey] as? String